restructure directory with utils, filters and run files
